Harden new parallel string_agg/array_agg regression test
Per buildfarm member mandrill, it seems that
max_parallel_workers_per_gather may not always be set to the default value
of 2 when the new test added in 1c900d9ff is executed.  Here let's just
explicitly set that to 2 so that the planner never opts to use more than
that many parallel workers.